+++ crates/driver/src/infra/api/routes/metrics.rs
@@ -1,13 +1,8 @@
-use prometheus::Encoder;
-
 pub(in crate::infra::api) fn metrics(app: axum::Router<()>) -> axum::Router<()> {
     app.route("/metrics", axum::routing::get(route))
 }
 
 async fn route() -> String {
     let registry = observe::metrics::get_registry();
-    let encoder = prometheus::TextEncoder::new();
-    let mut buffer = Vec::new();
-    encoder.encode(&registry.gather(), &mut buffer).unwrap();
-    String::from_utf8(buffer).unwrap()
+    observe::metrics::encode(registry)
 }
