+++ crates/solvers/src/infra/mod.rs
@@ -3,4 +3,4 @@ pub mod cli;
 pub mod config;
 pub mod contracts;
 pub mod dex;
-pub mod observe;
+pub mod metrics;
